Moved the TaskLogNosePlugin to ovirt
There's no need for it in the main package as all the nose actions are
in the ovirt one.

Fixes #99
